Fullscreen Loading
Put the loading of movies and tv to be 100% of the screen size, but still underneath the title bar. The loading screen background actually now lines up perfectly with the background of movies or that tv poster preview, so it feels a lot smoother; you'll have to build and run to see it since its not obvious in screenshots. Tested on Mac, seems perfectly responsive